refactor(engine): extract terminal-sync state machine from InlineRenderer
Phase 2, step 2. Engine (eye_declare_engine::Engine) now owns cursor
tracking, row accounting, the scrollback boundary, and the present/
reset/commit_scrolled/finalize machinery — it speaks Frames and bytes,
with no component-tree or crossterm dependency. InlineRenderer keeps
the tree concerns (push/state/focus/events, detect_committed, the tree
half of commit) and delegates terminal mechanics.

TestTerminal moves to eye_declare_engine::test_terminal behind the
test-util feature (vte becomes an optional engine dep), making the VTE
harness available to downstream users. Behavior-preserving: all 265
tests pass unchanged, including the scrollback regression suite.
